Unify usage funcs in logic pkg
